requirements.txt
pre-commit==2.19.0
flake8==4.0.1
pycodestyle==2.8.0
mypy==0.931
autopep8==1.6.0
alembic==1.7.5
amqp==5.0.6
aniso8601==9.0.0
apispec==4.3.0
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
boto3==1.16.15
botocore==1.19.15
celery==5.1.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
ConfigArgParse==1.4
cryptography==3.3.1
decorator==5.0.5
Flask==1.1.2
Flask-Admin==1.5.7
flask-apispec==0.11.0
Flask-BasicAuth==0.2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-expects-json==1.5.0
Flask-Injector==0.12.3
Flask-JWT-Extended==4.0.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
idna==2.10
injector==0.18.4
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.1
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.1.0
locust==1.4.3
Mako==1.1.3
MarkupSafe==2.0
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
more-itertools==8.6.0
msgpack==1.0.2
packaging==20.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.9.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python-editor==1.0.4
pytz==2020.5
# pyzmq==22.0.3
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
six==1.16.0
SQLAlchemy==1.3.24
timeago==1.0.15
toml==0.10.1
traitlets==5.1.0
typing-extensions==4.2.0
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
webargs==7.0.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.3.0
factory_boy==3.2.0
SQLAlchemy-Utils==0.37.8
Flask-APScheduler==1.12.2
dacite==1.6.0
pandas==1.4.2
sentry-sdk[flask]==1.5.6
pytest-timeout==2.0.1
newrelic==7.2.4.171
stripe==2.67.0
fabric==2.6.0
slack-bolt==1.13.0
google-auth==2.6.0
aiohttp==3.8.1
websocket==0.2.1
websockets==10.2
watchdog==0.9.0
Flask-SSE==1.0.0
Flask-Admin==1.5.7
Flask-Login==0.6.0
flask-celeryext==0.4.3
